+++ src/data.js
@@ -1,0 +1,22 @@
+export const abouts = [
+    {
+      name: 'Temtiope', 
+      about: 'my name is Temtiope and im a frontend developer',
+      location: 'lagos'
+    },
+    {
+      name: 'Victor', 
+      about: 'my name is Victor and im a frontend developer',
+      location: 'delta'
+    },
+    {
+      name: 'Richard', 
+      about: 'my name is Richard and im a frontend developer',
+      location: 'lagos'
+    },
+    {
+      name: 'Kayode', 
+      about: 'my name is Kayode and im a frontend developer',
+      location: 'lagos'
+    } ,
+]
