+++ src/studentlist/studentlist.css
@@ -1,0 +1,4 @@
+.main{
+  background-color: burlywood;
+  padding: 20px 50px;
+}
